+++ src/src-c/bjson-encode.h
@@ -89,7 +89,7 @@ BJSON_API bjson_status_t bjson_encoderReset(bjson_encodeCtx_t *ctx,
  */
 
 BJSON_API bjson_status_t bjson_encodeInteger(bjson_encodeCtx_t *ctx,
-                                             long long int value);
+                                             uint64_t int value);
 
 BJSON_API bjson_status_t bjson_encodeDouble(bjson_encodeCtx_t *ctx,
                                             double value);
